+++ mayan/apps/appearance/apps.py
@@ -4,8 +4,6 @@ from django.utils.translation import ugettext_lazy as _
 
 from mayan.apps.common.apps import MayanAppConfig
 
-from .dependencies import *  # NOQA
-
 
 class AppearanceApp(MayanAppConfig):
     name = 'mayan.apps.appearance'