+++ mayan/apps/metadata/apps.py
@@ -26,7 +26,6 @@ from mayan.apps.events.permissions import permission_events_view
 from mayan.apps.navigation.classes import SourceColumn
 
 from .classes import DocumentMetadataHelper
-from .dependencies import *  # NOQA
 from .events import (
     event_document_metadata_added, event_document_metadata_edited,
     event_document_metadata_removed, event_metadata_type_edited,